+++ frontend/src/pages/Manager/Manager.tsx
@@ -1,17 +1,17 @@
 import { useEffect, useState } from "react";
-import CashierNavbar from "../components/CashierNavbar";
-import { getUserAuth, UserInfo } from "./Login";
+import ManagerNavbar from "../../components/ManagerNavbar";
+import { getUserAuth, UserInfo } from "../Login";
 
-export default function Cashier() {
+export default function Manager() {
     const [userProfile, setUserProfile] = useState<UserInfo | undefined>(undefined);
 
     useEffect(() => {
-        getUserAuth('cashier')
+        getUserAuth('manager')
             .then(setUserProfile)
             .catch(console.error);
     }, [])
     
     return (userProfile &&
-        <CashierNavbar userInfo={userProfile} />
+        <ManagerNavbar userInfo={userProfile} />
     )
 }
